Rewrite stale vlm.* imports missed by the initial rename pass
10 CF-owned files were still importing from the old ``vlm`` module names
(``cosmos_framework.configs.base.defaults.vlm.*`` and
``cosmos_framework.model.vfm.vlm.qwen3_vl.utils``). After upstream i4
renamed the source subtree to ``reasoner``, these imports 500'd at
runtime because the target modules no longer exist. The initial rename
pass on this branch only caught the release-managed files; these CF-
owned files were missed.

Files rewritten (vlm -> reasoner):
- configs/base/config.py                                       (register_vlm)
- configs/base/defaults/open_source_dataloader.py              (create_qwen2_tokenizer_with_download)
- configs/base/experiment/sft/models/nano_model_config.py      (create_vlm_config, tokenizer)
- configs/base/experiment/sft/models/super_model_config.py     (same as nano)
- data/vfm/local_datasets/sft_dataset.py                       (tokenize_caption)
- inference/inference.py                                        (_SYSTEM_PROMPT_IMAGE_EDITING)
- inference/transfer.py                                         (_SYSTEM_PROMPT_TRANSFER)
- examples/integration/net_level.py                             (tokenize_caption)
- examples/integration/trainer_level_inference.py              (tokenize_caption)
- examples/integration/trainer_level_training.py               (tokenize_caption)

Verified by re-running ``test_launch_regression[vision_sft_nano]`` on
4 x GB200 — trains 10 iters and matches gb200 goldens exactly.
